requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.1
boto3==1.34.0
PyJWT==2.8.0 